refactor(main): add note for image and apt deb (#1307)

# Conflicts:
#	.goreleaser.yml
#	scripts/release/note.sh
