+++ cmd/lvscare/cmd/version.go
@@ -18,7 +18,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/labring/sealos/pkg/utils/version"
+	"github.com/labring/sealos/pkg/version"
 
 	"github.com/spf13/cobra"
 )
