+++ docker/lvscare/Dockerfile
@@ -1,4 +1,4 @@
-FROM scratch
+FROM alpine:latest
 
 COPY lvscare /usr/bin/lvscare
 ENTRYPOINT ["/usr/bin/lvscare"]
