+++ docker/sealos/Dockerfile
@@ -1,0 +1,4 @@
+FROM alpine:latest
+
+COPY sealos /usr/bin/sealos
+ENTRYPOINT ["/usr/bin/sealos"]
